Including KinFitter. Several parts commented but compiling.
